Cleaning up file structure
